removed unused API route
